Add files via upload
This project is done using HLS through C++, so all the source files were written in C++ language. Coming to the schematic, the terms of seconds and minutes are connected to the seven-segment driver, which gives the input to the seven-segment display. The refresh signal gives the input for the seven-segment driver.